# optimize setting up problem