--- a/src/npa_impl.jl
+++ b/src/npa_impl.jl
@@ -128,7 +128,7 @@ function sdp2jump(expr, ineqs;
           for (m, n) in size_as_pair.(ineqs)]
 
     Ids = (ineq[Id] for ineq in ineqs)
-    objective = (sum(LinearAlgebra.tr(s*m*z)
+    objective = (s*sum(LinearAlgebra.dot(m,z)
                      for (m, z) in zip(Ids, Zs))
                  + expr[Id])
 
@@ -143,7 +143,7 @@ function sdp2jump(expr, ineqs;
     for m in mons
         c = expr[m]
         Fs = (ineq[m] for ineq in ineqs)
-        tr_term = sum(LinearAlgebra.tr(F*Z)
+        tr_term = sum(LinearAlgebra.dot(F,Z)
                       for (F, Z) in zip(Fs, Zs))
 
         @constraint(model, tr_term + s*c == 0)