@araujoms 2024-07-04
And as I hoped, the tireless @blegat showed up to save the day: https://discourse.julialang.org/t/performance-of-primal-versus-dual-forms-in-jump/116633

Don't merge this yet, I'll implement his solution properly tomorrow.