@araujoms 2024-07-05
That's more like it. Now it's no longer the bottleneck:
```julia
Level 1:
  0.001012 seconds (3.58 k allocations: 119.250 KiB)
Level 2:
  0.000484 seconds (5.73 k allocations: 223.109 KiB)
Level 3:
  0.000681 seconds (11.13 k allocations: 525.609 KiB)
Level 4:
  0.000970 seconds (18.04 k allocations: 1.039 MiB)
Level 5:
  0.001976 seconds (28.31 k allocations: 2.019 MiB)
Level 6:
  0.002527 seconds (51.48 k allocations: 3.192 MiB)
Level 7:
  0.019090 seconds (71.75 k allocations: 5.575 MiB, 79.47% gc time)
Level 8:
  0.119416 seconds (96.67 k allocations: 10.214 MiB, 94.94% gc time)
Level 9:
  0.007954 seconds (127.38 k allocations: 13.196 MiB)
Level 10:
  0.012922 seconds (165.17 k allocations: 17.211 MiB, 15.39% gc time)
Level 11:
  0.016614 seconds (209.05 k allocations: 26.638 MiB, 12.03% gc time)
Level 12:
  0.020406 seconds (313.09 k allocations: 34.219 MiB)
Level 13:
  0.027814 seconds (386.22 k allocations: 51.232 MiB, 4.63% gc time)
Level 14:
  0.036658 seconds (468.35 k allocations: 61.807 MiB, 8.74% gc time)
Level 15:
  0.045647 seconds (552.55 k allocations: 74.515 MiB, 8.98% gc time)
Level 16:
  0.060822 seconds (644.48 k allocations: 103.631 MiB, 7.63% gc time)
```