@ewoodhead 2024-07-05
Thanks; I'll see if I can speed up the monomial multiplication & reduction, although I don't think I'll be able to work on it for a couple of weeks at least.