@ewoodhead 2024-07-05
I'll check but I'm not super optimistic. At the moment I get (on a slow VM):
```julia
julia> @time gamma = npa_moment(S, 16);
  5.718031 seconds (70.07 M allocations: 1.882 GiB, 6.82% gc time, 0.13% compilation time)
```
This is not dramatically slower than the old version (`git checkout bfad4e6 .`, from 29 Jul 2022):
```julia
julia> @time gamma = npa_moments(ops_at_level(S, 16));
  4.867666 seconds (61.98 M allocations: 1.460 GiB, 5.69% gc time)
```
For comparison, my old Lisp version (on the same VM) seems to be ten times as fast or more:
```
NPA-USER> (time
           (with-operators (A1 A2 B1 B2)
             (let ((ops (monomials-at-level (list A1 A2 B1 B2) 16)))
               (npa-moments ops))))
Evaluation took:
  0.346 seconds of real time
  0.346212 seconds of total run time (0.326470 user, 0.019742 system)
  100.00% CPU
  812,203,052 processor cycles
  230,122,336 bytes consed
```
So I'm doubtful if it's the algorithm that's the bottleneck.