@ewoodhead 2024-07-05
Well I can have a shot at profiling in a couple of weeks when I'm more likely to have the time. But in the end, if Julia isn't as fast as Lisp then there won't really be anything I can do about it.