@araujoms 2024-07-05
Lol there is no way Julia is the problem. Look at some benchmarks [here](https://julialang.org/benchmarks/). Julia is competitive with C and FORTRAN, which LISP is definitely not.