@ewoodhead 2024-07-06
No, compiled Common Lisp can get in the same ballpark with performance as C. See the benchmark results [here](https://www.norvig.com/python-lisp.html) or the results of [this study](https://dataverse.jpl.nasa.gov/dataset.xhtml?persistentId=hdl:2014/14462) for instance.